[feat/refactor] track amm swap params change events; major refactoring
