+++ dexi-core/overview.lua
@@ -1,0 +1,67 @@
+local sqlschema = require('dexi-core.sqlschema')
+
+local overview = {}
+
+function overview.getOverview(now, orderBy)
+  local orderByClause = "amm_discovered_at_ts DESC"
+
+  if orderBy == "volume" then
+    orderByClause = "volume DESC"
+  elseif orderBy == "transactions" then
+    orderByClause = "transactions DESC"
+  elseif orderBy == "market_cap" then
+    orderByClause = "market_cap DESC"
+  end
+
+  local stmt = db:prepare(string.format([[
+  WITH stats AS (
+    SELECT
+      amm_process,
+      COUNT(*) AS transactions,
+      SUM(volume) AS volume
+    FROM amm_transactions_view
+    WHERE created_at_ts >= :now - 86400
+    GROUP BY 1
+  ), current_prices AS (
+    SELECT
+      amm_process,
+      (SELECT price FROM amm_transactions_view WHERE amm_process = r.amm_process ORDER BY created_at_ts DESC LIMIT 1) AS current_price
+    FROM amm_registry r
+  )
+  SELECT
+    rank() OVER (ORDER BY t.total_supply * current_price DESC) AS market_cap_rank,
+    r.amm_name as amm_name,
+    r.amm_process as amm_process,
+    r.amm_token0 AS token0,
+    r.amm_token1 AS token1,
+    transactions,
+    volume,
+    t.token_name AS token_name,
+    t.total_supply AS total_supply,
+    t.fixed_supply AS fixed_supply,
+    t.total_supply * current_price AS market_cap,
+    (SELECT price FROM amm_transactions_view WHERE amm_process = r.amm_process ORDER BY created_at_ts DESC LIMIT 1) AS current_price,
+    (SELECT price FROM amm_transactions_view WHERE amm_process = r.amm_process AND created_at_ts <= :now - 300 ORDER BY created_at_ts DESC LIMIT 1) AS price_5m_ago,
+    (SELECT price FROM amm_transactions_view WHERE amm_process = r.amm_process AND created_at_ts <= :now - 3600 ORDER BY created_at_ts DESC LIMIT 1) AS price_1h_ago,
+    (SELECT price FROM amm_transactions_view WHERE amm_process = r.amm_process AND created_at_ts <= :now - 21600 ORDER BY created_at_ts DESC LIMIT 1) AS price_6h_ago,
+    (SELECT price FROM amm_transactions_view WHERE amm_process = r.amm_process AND created_at_ts <= :now - 86400 ORDER BY created_at_ts DESC LIMIT 1) AS price_24h_ago
+  FROM amm_registry r
+  LEFT JOIN stats s ON s.amm_process = r.amm_process
+  LEFT JOIN current_prices c ON c.amm_process = r.amm_process
+  LEFT JOIN token_registry t ON t.token_process = r.amm_token1
+  ORDER BY %s
+  LIMIT 100
+  ]], 'market_cap DESC'))
+
+  if not stmt then
+    error("Err: " .. db:errmsg())
+  end
+
+  stmt:bind_names({
+    now = now
+  })
+
+  return sqlschema.queryMany(stmt)
+end
+
+return overview
