+++ dexi-core/price-around.lua
@@ -1,0 +1,31 @@
+local sqlschema = require('dexi-core.sqlschema')
+
+local priceAround = {}
+
+function priceAround.findPriceAroundTimestamp(targetTimestampBefore, ammProcessId)
+  local stmt = db:prepare [[
+    SELECT price
+    FROM amm_transactions_view
+    WHERE created_at_ts <= :target_timestamp_before
+    AND amm_process = :amm_process_id
+    ORDER BY created_at_ts DESC
+    LIMIT 1;
+  ]]
+
+  if not stmt then
+    error("Failed to prepare SQL statement: " .. db:errmsg())
+  end
+
+  stmt:bind_names({
+    target_timestamp_before = targetTimestampBefore,
+    amm_process_id = ammProcessId
+  })
+
+
+  local row = sqlschema.queryOne(stmt)
+  local price = row and row.price or nil
+
+  return price
+end
+
+return priceAround
